Double precision for phred probability

+++ include/FastQ.cpp
@@ -69,9 +69,9 @@ namespace FastQ
     {
 
         int total_probability = 0;
-        float average_probability;
+        double average_probability;
 	int char_phred_qual;
-	float char_phred_prob;
+	double char_phred_prob;
 
         // Iterate through sequence, adding up quality probability of each nucleotide.
         for ( int i = 0; i < _length; i++ )
